// kanban-ui/src/app/model/card/card.ts
import {Dev} from '../dev/dev';

export class Card {
  id: number;
  name: string;
  type: string;
  author: Dev;
  worker: Dev;
  tester: Dev;
  timeCreated: string;
  timeTaken: string;
  timeDeadline: string;
  timeFactEndTime: string;
  status: string;
  direction: string;
  color: string;
  description: string;

}